frontend/admin/settings: refactor, confirm save-all, and disable autocomplete
